requirements.txt
playwright==1.16
pytest-playwright==0.2.0
pytest-html==3.1.1
pytest-rerunfailures==10.1
pytest-repeat==0.9.1
pytest-xdist==2.5.0
allure-pytest==2.9.43
pytest-ordering==0.6
pytest-assume==2.4.3
pytest-sugar==0.9.4
PyMySQL==1.0.2

pytest~=6.2.5
PyYAML==6.0